CNJR-3040 Self Updating Summon Formula
This is a homebrew formula for installing summon that doesn't
have a hardcoded version - it finds the latest release from GitHub
and provides that.

The versioning is maintained, so brew will know the current version
that is installed and when there is an update available.

Installing old versions won't work, but that doesn't work currently
either AFAICT.

The motivation for this is to the able to automatically release
projects that publish to homebrew and not have to push commits
to this repo for each release.