CNJR-9177: Fix install from brew
This was failing as the hash asset was not listed by the api call that github-update was using. Before this commit we were requesting all releases for a repo from the github api, then finding the latest (zero index), extracting the assets url, then requesting assets from the assets url.

This commit simplifies that process by requesting /releases/latest. This gives us the latest release directly, it also includes the assets list so we don't need a second request to fetch assets.