Set HOMEBREW_DEVELOPER in test environment
Homebrew requires formulae to be part of a tap as of homebrew 4.6.4. This broke
our test environment because we install the formulae into our test env using
`brew install --build-from-source` without using `tap`. The solution of setting
HOMEBREW_DEVELOPER can be found in the [PR for this change](https://github.com/Homebrew/brew/pull/20414).